component refactor search
