+++ src/components/student-portal/all-classes/AllClasses.page.jsx
@@ -3,6 +3,7 @@ import React, { useEffect, useState } from 'react';
 import DisplayAllClasses from './DisplayAllClasses.component';
 import filterPrograms from '../../../utils/search-filter';
 import { Box, Stack, TextField } from '@mui/material';
+import SearchFilters from '../../search-filters/SearchFilters.component';
 
 export default function StudentPortalAllClasses() {
     const [allClasses, setAllClasses] = useState([]);
@@ -28,41 +29,7 @@ export default function StudentPortalAllClasses() {
                 <p>Here you can see all the classes available.</p>
             </div>
 
-            <div className="bg-blue-200 w-full md:w-4/6 shadow-lg h-5/6 gap-4 p-2 md:p-4 border rounded-lg ">
-                <Box
-                    component="fieldset"
-                    sx={{
-                        display: 'flex',
-                        flexDirection: 'column',
-
-                        '> legend': {
-                            textAlign: 'left',
-                            fontWeight: 'bold',
-                            marginBottom: '.5rem'
-                        },
-
-                        'div > div': {
-                            background: 'white'
-                        }
-                    }}
-                >
-                    <Stack spacing={2} direction="row">
-                        <TextField
-                            color="black"
-                            id="search-by-name"
-                            label="Class Name"
-                            value={searchByName}
-                            onChange={typedByUser => {
-                                setSearchByName(
-                                    typedByUser.target.value.toLocaleLowerCase()
-                                );
-                            }}
-                            sx={{ flex: 1 }}
-                        />
-                    </Stack>
-                </Box>
-            </div>
-
+            <SearchFilters searchByName={searchByName} setSearchByName={setSearchByName}/>
             <DisplayAllClasses allClasses={filteredClasses} isAdmin={false} />
         </>
     );
